@@ -5,24 +5,16 @@
  */
 package io.mosip.esignet.mock.identitysystem.controller;
 
-import io.mosip.esignet.mock.identitysystem.dto.Error;
 import io.mosip.esignet.mock.identitysystem.dto.PartnerDto;
 import io.mosip.esignet.mock.identitysystem.dto.RequestWrapper;
 import io.mosip.esignet.mock.identitysystem.dto.ResponseWrapper;
 import io.mosip.esignet.mock.identitysystem.service.PartnerService;
 import io.mosip.esignet.mock.identitysystem.util.HelperUtil;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.ConstraintViolation;
-import javax.validation.ConstraintViolationException;
-import javax.validation.Valid;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
+import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("/")
@@ -41,24 +33,4 @@ public class PartnerController {
         return response;
     }
 
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity handleException(ConstraintViolationException ex) {
-        List<io.mosip.esignet.mock.identitysystem.dto.Error> errors = new ArrayList<>();
-        if(ex != null && ex instanceof ConstraintViolationException) {
-            Set<ConstraintViolation<?>> violations = ((ConstraintViolationException) ex).getConstraintViolations();
-            for(ConstraintViolation<?> cv : violations) {
-                errors.add(new Error(cv.getMessage(), cv.getPropertyPath().toString() + ": " + cv.getMessage()));
-            }
-            return new ResponseEntity<ResponseWrapper>(getResponseWrapper(errors), HttpStatus.OK);
-        }
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                .body(errors);
-    }
-
-    private ResponseWrapper getResponseWrapper(List<Error> errors) {
-        ResponseWrapper responseWrapper = new ResponseWrapper<>();
-        responseWrapper.setResponseTime(HelperUtil.getCurrentUTCDateTime());
-        responseWrapper.setErrors(errors);
-        return responseWrapper;
-    }
 }
